+++ apps/routerconsole/java/src/net/i2p/router/web/UpdateHandler.java
@@ -236,7 +236,7 @@ public class UpdateHandler {
             StringBuilder buf = new StringBuilder(64);
             buf.append("<b>").append(_("Updating")).append("</b> ");
             double pct = ((double)alreadyTransferred + (double)currentWrite) /
-                         ((double)alreadyTransferred + (double)currentWrite + (double)bytesRemaining);
+                         ((double)alreadyTransferred + (double)currentWrite + bytesRemaining);
             synchronized (_pct) {
                 buf.append(_pct.format(pct));
             }
